requirements.txt
Flask
tensorflow
numpy
librosa
opencv-python
torch
torchvision
Pillow
